requirements.txt
flask
torch
torchvision
pillow
gdown
gunicorn